+++ src/FileReader/ReaderFactory.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace TomPHP\ConfigServiceProvider;
+namespace TomPHP\ConfigServiceProvider\FileReader;
 
 use TomPHP\ConfigServiceProvider\Exception\UnknownFileTypeException;
+use TomPHP\ConfigServiceProvider\FileReader\FileReader;
 
 final class ReaderFactory
 {